+++ lib/fluent_ui.dart
@@ -27,7 +27,7 @@ export 'src/navigation/route.dart';
 
 export 'src/layout/page.dart';
 
-export 'src/controls/inputs/hover_button.dart';
+export 'src/controls/utils/hover_button.dart';
 export 'src/controls/inputs/button.dart';
 export 'src/controls/inputs/checkbox.dart';
 export 'src/controls/inputs/icon_button.dart';